Do not show empty security group alert if error
